sync dependencies from OPM
We want to get rid of OPM package, so let's first sync the list of
modules with OPM in puppet-tripleo, so we can rely on puppet-tripleo to
deploy all Puppet modules that TripleO requires.

Also order by alphabetic so easier to manage.

Also it requires this patch for lint: https://review.openstack.org/510221

Depends-On: I856eb88ae8e9ffe88879a8c26204a83d5d7c956c
Change-Id: Ibbc3ea5950f610c24137b51682aef2fbd8b8d04b