Make sure iptables-services gets pulled in on rhel8
Currently puppetlabs-firewall only pulls in the nftables package.
On rhel8 We want to pull in the iptables-services as puppet-firewall
needs the /usr/libexec/iptables/iptables[6].init scripts in order
to persist the iptables rules. We're adding this to puppet-tripleo as
this is specific to the way we manage iptables rules in tripleo. We are
already pulling this in the images but we need to manage this in the
containers as well.

Change-Id: I7192dd5b467dea765e21923c05ae4fb1436f56fa