Remove puppet-ceph dependency from puppet-tripleo
When OSP is upgraded from Train there is no need to
keep puppet-ceph package anymore because ceph-ansible
is used for Ceph related deployments. Same story for
fresh deployments.
The TripleO review [1] has an additional post_upgrade
task built with the purpose of removing puppet-ceph
to clean up the system.
This change just removes the puppet-ceph pkg as a dep.

[1] https://review.opendev.org/#/c/719050

Change-Id: I12bd70ded09c9ec7ea6afda178d4f0e192cb7ef9